Moved some common functions to the utils module
